🐛 fix(content): move author-slug helpers out of the fs-backed posts module
PostAuthors ("use client") imported generateAuthorSlug from
lib/content/posts/posts.ts, which transitively imports fs via
lib/content/content.ts — Turbopack correctly refused to bundle fs for
the browser. The pure id<->slug helpers now live in a standalone
lib/content/authors/author-slug.ts with no fs dependency; posts.ts
re-exports them for its existing server-side consumers.
