+++ src/lib/content/posts/posts.test.ts
@@ -1,11 +1,8 @@
 import { describe, it, expect } from "vitest";
 import {
     aggregateAuthorsWithPosts,
-    authorIdToSlug,
-    authorSlugToId,
     filterPostsForAuthor,
     findAuthorWithPostsBySlug,
-    generateAuthorSlug,
     rankReadNextPosts,
 } from "./posts";
 import { Content } from "@/types/content/content";
@@ -128,33 +125,6 @@ describe("rankReadNextPosts", () => {
             expect(result[0].slug.formatted).toBe("post-a");
             expect(result[1].slug.formatted).toBe("post-b");
         });
-    });
-});
-
-describe("authorIdToSlug", () => {
-    it("replaces underscores with hyphens", () => {
-        expect(authorIdToSlug("fabrizio_duroni")).toBe("fabrizio-duroni");
-    });
-
-    it("replaces every underscore for multi-word keys", () => {
-        expect(authorIdToSlug("marco_de_lucchi")).toBe("marco-de-lucchi");
-    });
-});
-
-describe("authorSlugToId", () => {
-    it("replaces hyphens with underscores", () => {
-        expect(authorSlugToId("fabrizio-duroni")).toBe("fabrizio_duroni");
-    });
-
-    it("round-trips with authorIdToSlug", () => {
-        const id = "marco_de_lucchi";
-        expect(authorSlugToId(authorIdToSlug(id))).toBe(id);
-    });
-});
-
-describe("generateAuthorSlug", () => {
-    it("builds the author detail href from the hyphenated slug", () => {
-        expect(generateAuthorSlug("fabrizio_duroni")).toBe("/blog/author/fabrizio-duroni");
     });
 });
